Remove redundant cleanups in test_volume_backup
The BaseVolumeTest class contains some helper methods that do their
own cleanup steps within them. However, several of the tests in
VolumesBackupsTest were also doing identical cleanup.

Occasionally, this results in a failure to cleanup in the gate [1]
because a test that doesn't wait for resource deletion [2] after it
cleans up a resource can leave a resource in a mid-deletion state that
the base class cleanup will fail with a 400 when *it* tries to delete
the same resource.

This was noticed on the Ocata branch but it looks like the same issues
exist on master in the plugin, so I'm starting with the fix here.

Closes-Bug: #1774684

[1] http://logs.openstack.org/53/570653/1/check/legacy-tempest-dsvm-full-devstack-plugin-ceph/c5d03e7/job-output.txt.gz#_2018-05-30_04_41_10_689262
[2] https://github.com/openstack/cinder-tempest-plugin/blob/28456a1/cinder_tempest_plugin/api/volume/test_volume_backup.py#L64

Change-Id: I2a29daf56f8327fce405e7969767dc9993849f19