GET backup before asserting 'volume_id' and 'snapshot_id'
Recently, commit c96757b275a77a745c6c80947bdd0bcb2de7eefd was merged
to remove redundant cleanups in volume backup tests. But there was a
bug in the change that assumed 'volume_id' and 'snapshot_id' are
returned from the backup create API when they are not.

We need to GET the backup in the API in order to obtain the 'volume_id'
and 'snapshot_id' for the asserts.

Closes-Bug: #1796207

Change-Id: I640fe41ec6049bfd980dfcbb6cae95a9a3e97b66