Add Snapshot data integrity test
This patch adds data integrity tests for snapshot with the following
procedure :

1) create a volume from image
2) Boot an instance from the volume
3) create file on vm and write data into it
4) create first snapshot
5) repeat 3
5) create second snapshot
6) repeat 3
7) create third snapshot

Now restore the snapshots one by one into volume, create instances
from it and check the number of files and file content at each
point snapshot was created.

We are also temporarily making a copy of tempest scenario manager until
the actual manager is available for import in tempest plugins.

Change-Id: I5c5ff6f996dd39b52fada82f3938b628e58b0a2c