Create new Ceph job with c-bak service enabled
Currently backup service is disable in Ceph jobs.
In order to solve this cinder-tempest-plugin-cbak-ceph is
defined with the c-bak service enabled.

As a matter of fact the c-bak is redundant incase of lvm job
because the parent job already has enabled the service.

In addition, Tempest provides a blacklist file so that only
storage-relevant tests are executed. Use it instead of the
custom blacklist, which is removed in this patch.

This means that some tests which were previously skipped are
now executed.

If backup tests don't work with the LVM backend, they can be
disabled through the volume_feature_enabled.backup configuration
key.

Co-authored-by: Luigi Toscano <ltoscano@redhat.com>
Change-Id: Id809c37704d4ee741df4333ca01ddd4ec22f33b5