Disable siblings for the cindelib functional tests
Follow the same pattern of the other jobs which deploy devstack
but runs also functional tests (see devtack-tox-functional
and devstack-tox-functional-consumer).

Without this, the older branches where cinderlib functional tests
use Python 2 (stein, and soon train when that is branched) fails
because they try to install packages part of which are Python 3
only.

Change-Id: Ibd2862f30d4c451118fb96d1f56a09fd66b2d75a