Use the embedded scenario manager instead of tempest one
While waiting for tempest scenario manager to stabilize its API,
a copy of it has been imported, so make sure to use it and to not
rely on tempest's one (whose API is changing during the stabilization
process, so it may break anytime).

Just one file needs to change its imports.
At the same time, fix the import order for another file.

Change-Id: I29c4784d59151948778cad57b7db56a368ed821d