Ceph jobs: re-enable a few c-bak tests, add more jobs
Those tests were disabled when ceph was upgraded
to a newer release (I51c90e592070b99422e692d5e9e97083d93146e8)
because they were failing.
But it seems they are working now, and it is not not trivial
at this point to understand whether something was fixed
in ceph or in cinder, so let's re-enable them.

Also, add a few additional (initially experimental) jobs
to double-check the ceph/c-bak scenario against
the stable branches.

This reverts commit 9d626d0de5ebddbe359b3f3bf0dae4c6e45efaad
and adds more functionalities.

Change-Id: I033e0dc04f342e27d6266ad412d3cb256787c0e6