Add command to fix/generate idempotent IDs
This patch adds a "uuidgen" command that will fix all issues
related to idempotent IDs in our tests. Also modifies pep8 to
detect these issues.
This command already exists in tempest and is required for cinder
tempest plugin as we had undetected issues in our tests that
got merged which are also fixed in this patch. The issues are as
follows:

1) Missing idempotent IDs in test_capabilities
2) Duplicate idempotent IDs in the following backup tests
    a) test_backup_create_and_restore_to_an_existing_volume
    b) test_incr_backup_create_and_restore_to_an_existing_volume

Change-Id: I6b34fd67af4c302ff17244143506d8c8a8247eb6