Moving API microversion fixture in resource_setup
API microversion fixture is used to set the API
microversion on service client and reset it back
after test execution. That is done in setUp()
method and due to that API microversion is not
set on service client during resource_setup()
method and any API call will be requested on base
version.

To sync the complete test execution of test class
with requested API microversion, this commit moves
the API microversion setting in resource_setup()
method.

Related-Bug: #1881863
Change-Id: Id26c73e4f0554770cd13d2b8840fe430448cc87f