Multiattach: Fix implicit requirement for encryption support
This test requires encrypted volume support but doesn't check
the tempest option for whether that support is enabled.

(This currently would mostly affect the NFS driver.)

For this test, encryption is not strictly required, so use it
when available to get a more interesting test scenario, but just
use regular volumes when the option is disabled.

This change maintains the original behavior that three volume
types are being used (default, multiattach, and encrypted/other),
but it isn't clear if the additional volume types really add
much of value to the test scenario.

Closes-Bug: #2043830
Change-Id: If41efc6ebd07c75d14b7d0cc5be183e9da7c8cfa