Conditionallly skip manage snapshot tests
The cinder manage snapshots commands are not supported by all drivers,
so the RBAC tests should be conditional to the driver supporting it.

The unmanage snapshot RBAC tests can still be run, because the failure
should happen at the API level, so it doesn't matter if the driver
supports it or not.

Change-Id: I2a65f8655a7418a9099e8d7593bd7d7009f22d9d