zuul: move the storage exclude list to the lvm-lio abstract job
Probably for historical reason this was not added to the "eldest" job
where it was needed. The result was that the lvm-nvmet-barbican job
did not use the exclude list, thus running more tests than needed.
A very old variant (probably not in use anymore) requires the old
version of the exclude list variable, and that's left there,
and the old variable being inherited shouldn't be a problem.

Change-Id: Ie576bc660df829288ebdd4bf952bad3c19b2bbcd