Add 'encryption_id' parameter to enc type calls
Change Ieca29000b575 in tempest revised the update and delete
encryption type methods for the encryption_type_client to include
the encryption_id, as is specified in the Block Storage API ref.
Thus we need to update our usage of the encryption_type_client to
pass the encryption_id parameter.

Partial-bug: #1835186

Change-Id: I14d2eb8ec8f4343464710124df0c108464bece1b